requirements.txt
argcomplete==1.12.3
asttokens==2.0.5
attrdict==2.0.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
base58==2.1.1
black==22.1.1.dev41+gf239d22
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
cytoolz==0.11.2
dataclasses==0.8
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
eth-abi==1.3.0
eth-account==0.3.0
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
ethereum-dasm==0.1.4
ethereum-etl==1.10.1
evmdasm==0.1.8
executing==0.8.2
Flask==2.0.3
google-api-core==1.31.5
google-auth==1.35.0
google-cloud-core==1.7.2
google-cloud-pubsub==2.1.0
google-cloud-storage==1.33.0
google-crc32c==1.3.0
google-resumable-media==1.3.3
googleapis-common-protos==1.55.0
grpc-google-iam-v1==0.12.3
grpcio==1.44.0
hexbytes==0.2.2
idna==3.3
ipykernel==6.9.1
ipython==8.0.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.1
jupyter-client==7.1.2
jupyter-core==4.9.2
kafka-python==2.0.2
libcst==0.3.21
lru-dict==1.1.7
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
numpy==1.22.3
packaging==21.3
pandas==1.4.1
parsimonious==0.8.1
parso==0.8.3
pathspec==0.9.0
pg8000==1.13.2
pickleshare==0.7.5
pip==21.2.4
pipx==0.16.4
platformdirs==2.5.1
prompt-toolkit==3.0.27
proto-plus==1.20.3
protobuf==3.19.4
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.14.1
Pygments==2.11.2
pyparsing==3.0.6
pypiwin32==223
python-dateutil==2.8.2
python-dotenv==0.20.0
pyTigerDriver==1.0.14
pyTigerGraph==0.0.9.9.2
pytz==2021.3
pywin32==303
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
rlp==2.0.1
rsa==4.8
scramp==1.1.0
setuptools==58.0.4
six==1.16.0
SQLAlchemy==1.3.13
stack-data==0.2.0
tabulate==0.8.9
timeout-decorator==0.4.1
tomli==2.0.1
toolz==0.11.2
tornado==6.1
traitlets==5.1.1
typed-ast==1.5.2
typing_extensions==4.1.1
typing-inspect==0.7.1
urllib3==1.26.8
userpath==1.7.0
validators==0.18.2
wcwidth==0.2.5
web3==4.7.2
websockets==6.0
Werkzeug==2.0.3
wheel==0.37.1
wincertstore==0.2
